Refactor and build of app
